Remove @author tags per Drupal best practices.
First, multiple people have contributed to this code. Second, this is to
ignore warnings from PHP_CodeSniffer. With recent versions of
PHP_CodeSniffer, it does not seem possible to ignore these warnings
without disabling PHP_CodeSniffer for the entire comment block. So it is
easiest to just remove the offending @author tags.